requirements.txt
cryptography
aiohttp
aiofiles
paho-mqtt
python-dotenv
pyyaml
prometheus-client
fastapi
uvicorn